pyproject.toml
[project]
name = "digital-forensics-detector-api"
version = "1.0.0"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi==0.115.5",
  "uvicorn[standard]==0.32.1",
  "python-multipart==0.0.18",
  "Pillow==11.0.0",
]
